# api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: developers@givechariot.com
    name: Chariot Development Team
    url: https://www.givechariot.com/contact
  description: The Chariot Payments REST API.
  termsOfService: https://www.givechariot.com/legal-nonprofit
  title: Chariot Payments API
  version: v1
servers:
- description: Staging
  url: https://devapi.givechariot.com
- description: Sandbox
  url: https://sandboxapi.givechariot.com
- description: Production
  url: https://api.givechariot.com
security:
- bearerAuth: []
paths:
...
  /recipients:
    get:
      description: |
        List recipients.
        This endpoint is used to list recipients.
      operationId: listRecipients
      parameters:
      - description: The Employer Identification Number (EIN) for the recipient.
        explode: true
        in: query
        name: ein
        required: false
        schema:
          type: string
        style: form
      - description: Limit the size of the list that is returned. The default (and
          maximum) is 100 objects.
        explode: true
        in: query
        name: limit
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: "The cursor to use for pagination. If not set, the first page\
          \ of results will be returned."
        explode: true
        in: query
        name: cursor
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipientList'
          description: The recipients were retrieved
        "400":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 400
                error: Bad Request
                message: Expected an API key to be provided in the header `x-chariot-api-key`
              schema:
                $ref: '#/components/schemas/ApiError'
          description: The request is invalid or contains invalid parameters
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "401":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 401
                error: Unauthorized
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Unauthorized. The request is missing the security (OAuth2 Bearer
            token) requirements and the server is unable to verify the identify of
            the caller.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "403":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 403
                error: Forbidden
                message: User is not allowed to access this resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Access denied
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
      summary: List recipients
      tags:
      - recipients
    post:
      description: |
        Create a recipient.
        This endpoint is used to create a recipient.
      operationId: createRecipient
      requestBody:
        $ref: '#/components/requestBodies/CreateRecipientRequest'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipient'
          description: The recipient was created
          headers:
            Location:
              $ref: '#/components/headers/Location'
        "400":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 400
                error: Bad Request
                message: Expected an API key to be provided in the header `x-chariot-api-key`
              schema:
                $ref: '#/components/schemas/ApiError'
          description: The request is invalid or contains invalid parameters
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "401":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 401
                error: Unauthorized
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Unauthorized. The request is missing the security (OAuth2 Bearer
            token) requirements and the server is unable to verify the identify of
            the caller.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "409":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 409
                error: Conflict
                message: Request conflicts with an already existing resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Resource Conflicts
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
      summary: Create a recipient
      tags:
      - recipients
  /recipients/{id}:
    get:
      description: |
        Get a recipient by its unique identifier.
        This endpoint is used to retrieve a recipient by its unique identifier.
      operationId: getRecipient
      parameters:
      - description: The unique identifier for the recipient
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipient'
          description: The recipient was retrieved
        "400":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 400
                error: Bad Request
                message: Expected an API key to be provided in the header `x-chariot-api-key`
              schema:
                $ref: '#/components/schemas/ApiError'
          description: The request is invalid or contains invalid parameters
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "401":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 401
                error: Unauthorized
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Unauthorized. The request is missing the security (OAuth2 Bearer
            token) requirements and the server is unable to verify the identify of
            the caller.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "403":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 403
                error: Forbidden
                message: User is not allowed to access this resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Access denied
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "404":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 404
                error: Not Found
                message: Resource is not found
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Resource Not Found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "409":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 409
                error: Conflict
                message: Request conflicts with an already existing resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Resource Conflicts
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
      summary: Retrieve a recipient
      tags:
      - recipients
...
components:
  headers:
    X-Request-Id:
      description: The unique identifier for the request
      explode: false
      schema:
        type: string
      style: simple
...
    Location:
      description: The URI reference of the created object.
      explode: false
      schema:
        format: uri
        type: string
      style: simple
  requestBodies:
    CreateRecipientRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/createRecipient_request'
      description: The request body for creating a recipient
      required: true
...
  schemas:
...
    Recipient:
      description: A recipient is a verified nonprofit organization account that can
        receive payments.
      example:
        name: Nonprofit
        created_at: 2000-01-23T04:56:07.000+00:00
        id: charity_658744hjksd23j
        ein: "123456789"
        primary: true
        status: verified
      properties:
        id:
          description: The unique identifier for the recipient
          example: charity_658744hjksd23j
          readOnly: true
          type: string
        name:
          description: The name of the recipient
          example: Nonprofit
          type: string
        ein:
          description: The Employer Identification Number (EIN) for the recipient.
          example: "123456789"
          type: string
        primary:
          description: |
            Indicates whether the recipient is the primary recipient for the EIN.
            Only one recipient can be the primary recipient for an EIN.
          example: true
          type: boolean
        status:
          description: The status of the recipient
          enum:
          - unverified
          - verified
          example: verified
          readOnly: true
          type: string
        created_at:
          description: The date and time the recipient was created
          format: date-time
          readOnly: true
          type: string
      type: object
    RecipientList:
      description: A list of recipients
      example:
        data:
        - name: Nonprofit
          created_at: 2000-01-23T04:56:07.000+00:00
          id: charity_658744hjksd23j
          ein: "123456789"
          primary: true
          status: verified
        - name: Nonprofit
          created_at: 2000-01-23T04:56:07.000+00:00
          id: charity_658744hjksd23j
          ein: "123456789"
          primary: true
          status: verified
        paging:
          cursors:
            before: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
            after: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
          total: 0
      properties:
        data:
          description: The list of recipients
          items:
            $ref: '#/components/schemas/Recipient'
          type: array
        paging:
          $ref: '#/components/schemas/RecipientList_paging'
    RecipientBankAddress:
      description: The bank address for a recipient
      properties:
        account_number:
          description: The account number
          example: "1234567890"
          type: string
        routing_number:
          description: The American Bankers' Association (ABA) Routing Transit Number
            (RTN).
          example: "123456789"
          type: string
      required:
      - account_number
      - routing_number
      type: object
...
    ApiError:
      description: An API error
      properties:
        timestamp:
          description: The date and time the error occurred
          format: date-time
          readOnly: true
          type: string
        status:
          description: The HTTP status code
          readOnly: true
          type: integer
        error:
          description: The HTTP status message
          readOnly: true
          type: string
        message:
          description: The error message
          readOnly: true
          type: string
      type: object
    createRecipient_request:
      properties:
        name:
          description: The name of the recipient
          example: Chariot Giving
          type: string
        ein:
          description: The Employer Identification Number (EIN) for the recipient.
          example: "123456789"
          type: string
        primary:
          description: |
            Indicates whether the recipient is the primary recipient for the EIN.
            Only one recipient can be the primary recipient for an EIN.
          example: true
          type: boolean
        bank_address:
          $ref: '#/components/schemas/RecipientBankAddress'
      type: object
    Pagination_cursors:
      description: The cursors for the next and previous pages
      example:
        before: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
        after: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
      properties:
        before:
          description: The cursor for the previous page
          example: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
          type: string
        after:
          description: The cursor for the next page
          example: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
          type: string
      type: object
    RecipientList_paging:
      description: The paging information
      example:
        cursors:
          before: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
          after: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
        total: 0
      properties:
        cursors:
          $ref: '#/components/schemas/Pagination_cursors'
        total:
          type: integer
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: Opaque
      scheme: bearer
      type: http